feat: restructure project pages
