+++ app/routes/_auth.tsx
@@ -4,7 +4,7 @@ type Props = {};
 
 function LayoutRoute(props: Props) {
   return (
-    <main className="flex min-h-screen items-center justify-center">
+    <main className="flex min-h-screen w-full items-center justify-center">
       <Outlet />
     </main>
   );
